+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @uploadthing/nuxt
 
+## 6.5.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @uploadthing/vue@6.5.3
+
+## 6.5.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @uploadthing/vue@6.5.2
+
 ## 6.5.2
 
 ### Patch Changes